Changes new room restrictions certain teachers input and error flagging
This commit changes the default form mode for the input in both the now and
future restrictions for certain teachers to start by prompting for the first
teacher rather than prompting the user whether or not to add a chip
because if the user would like to have a restriction, there must be a
teacher which will approve the pass for the restriction, so it changes that.

This commit also makes it so that the "Restricion for now" and "Restriction for
future" become red if the user has not selected any teachers but tries to save
the pass. More things can become red, just let me know, and I'll make that change.