Adds tooltip for new room/folder save while still showing errors
This commit adjusts the way errors are shown when a field isn't filled
out correctly in the new room/folder forms. Instead of having a
continuously active button, this commit makes it so buttons can have an
optional event emitter even when they are disabled, which can be
seperate from the normal event emitter. Showing errors is the event
fired when the button is disabled. The button is disabled when the form
is incomplete.

Further, this commit changes the tooltip so that if the form is
incomplete, the tooltip will show which fields are not filled out
properly. Note that, the previous tooltip is still activatable under its
previously set conditions. The default is no tooltip.