SP-1119: Frontend acquisition forms including quoterequest and predemo
This commit adds the forms module which has a predemo form, and a
quote request form. This commit also includes two fields including
the colored checkbox field (essentially the airtable checkboxes),
and list schools field where the user can list schools. This commit
makes use of @types/googlemaps and @agm/core which use a api key
which should be changed.
This commit does not ineract with the backend because that is not
public, but it does format its data in such a way where it would be
easy to add, once the backend changes are approved.