SPB-424: Create offline bar to show users when websocket is disconnected, and allow the user to reconnect.
The polling service now saves a websocket, so old websockets which might be slow can be closed. The raw messages and events were seperated so that when a websocket switches, there don't need to be updates to either of the subjects, or subscribers. The user can try to reconnect the websocket, but it will not go through if a reconnection has been attempted in the last two seconds.