Fixed pass limit filter showing only some results
As we paginate accounts for the table (as we should)
our filter callback (used by Individual Pass Limit button)
only filters on the first page.
Since our nextPage pagination request is based off a large number of
rows in the table before requesting another page, this
means that if the filter didn't return a large number of items, the
filter would be an incomplete query from the user's perspective.
As a fix, if there is a small number of rows returned by the
filter, we will now keep making requests until it is of large size
or the end of the table is reached.